refactor: Move integration tests to dedicated package
Consolidates all integration tests into domain.integration package:
- M5IntegrationTest
- M6bStrategyIntegrationTest
- M6EndToEndIntegrationTest
- ContributionIntegrationTest
- PhaseOutIntegrationTest
