+++ src/test/java/io/github/xmljim/retirement/domain/integration/M5IntegrationTest.java
@@ -1,4 +1,4 @@
-package io.github.xmljim.retirement.domain;
+package io.github.xmljim.retirement.domain.integration;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
